overall changes

+++ src/user/user.controller.ts
@@ -3,6 +3,8 @@ import {
   Controller,
   Delete,
   Get,
+  HttpException,
+  HttpStatus,
   Param,
   Patch,
   Post,
@@ -11,7 +13,7 @@ import { CreateUserDto } from './DTO/createUserDto';
 import { EditPasswordDto } from './DTO/editPasswordDto';
 import { UpdateUserDto } from './DTO/updateUserDto';
 import { UserService } from './user.service';
-
+import { isValidObjectId, ObjectId } from 'mongoose';
 @Controller('user')
 export class UserController {
   constructor(private readonly userService: UserService) {}
@@ -28,16 +30,31 @@ export class UserController {
 
   @Get('/:id')
   async getUser(@Param('id') id) {
+    if (!isValidObjectId(id))
+      throw new HttpException(
+        'Not a valid mongoDb id!',
+        HttpStatus.PRECONDITION_FAILED,
+      );
     return await this.userService.get(id);
   }
 
   @Delete('/:id')
   async deleteUserById(@Param('id') id) {
+    if (!isValidObjectId(id))
+      throw new HttpException(
+        'Not a valid mongoDb id!',
+        HttpStatus.PRECONDITION_FAILED,
+      );
     return await this.userService.deleteById(id);
   }
 
   @Post('/:id')
   async updateUserById(@Param('id') id, @Body() updatedUser: UpdateUserDto) {
+    if (!isValidObjectId(id))
+      throw new HttpException(
+        'Not a valid mongoDb id!',
+        HttpStatus.PRECONDITION_FAILED,
+      );
     return await this.userService.updateUser(id, updatedUser);
   }
 
@@ -46,6 +63,11 @@ export class UserController {
     @Param('id') id,
     @Body() newPassword: EditPasswordDto,
   ) {
+    if (!isValidObjectId(id))
+      throw new HttpException(
+        'Not a valid mongoDb id!',
+        HttpStatus.PRECONDITION_FAILED,
+      );
     return await this.userService.updatePassword(id, newPassword);
   }
 }
